# Cache l2 tables